Friends and Reaction controller functions

+++ controllers/thoughtController.js
@@ -12,6 +12,11 @@ module.exports = {
   },
   async getSingleThought(req, res) {
     try {
+      if(!ObjectId.isValid(req.body.thoughtId)){
+        return res.status(404).json({
+          message: 'ID is invalid!',
+        });
+      }
       const thought = await Thought.findOne({ _id: req.params.thoughtId })
       if (!thought) {
         return res.status(404).json({ message: 'No thought with that ID' });
@@ -51,6 +56,11 @@ module.exports = {
   },
   async updateThought(req, res) {
     try {
+      if(!ObjectId.isValid(req.params.thoughtId)){
+        return res.status(404).json({
+          message: 'ID is invalid!',
+        });
+      }
       const thought = await Thought.findOneAndUpdate(
         { _id: req.params.thoughtId },
         { $set: req.body },
@@ -69,6 +79,11 @@ module.exports = {
   },
   async deleteThought(req, res) {
     try {
+      if(!ObjectId.isValid(req.params.thoughtId)){
+        return res.status(404).json({
+          message: 'ID is invalid!',
+        });
+      }
       const thought = await Thought.findOneAndRemove({ _id: req.params.thoughtId });
 
       if (!thought) {
@@ -92,9 +107,14 @@ module.exports = {
       res.status(500).json(err);
     }
   },
-  // Add a thought response
+  // Add a thought reaction
   async addThoughtReaction(req, res) {
     try {
+      if(!ObjectId.isValid(req.params.thoughtId)){
+        return res.status(404).json({
+          message: 'ID is invalid!',
+        });
+      }
       const thought = await Thought.findOneAndUpdate(
         { _id: req.params.thoughtId },
         { $addToSet: { reactions: req.body } },
